src/backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Kafka
confluent-kafka==2.3.0
avro-python3==1.10.2
fastavro==1.9.0

# ML
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
joblib==1.3.2
shap==0.44.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# HTTP client
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# AWS Lambda adapter
mangum==0.17.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
structlog==24.1.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
pylint==3.0.3
mypy==1.8.0

